Transition from `JSON::Matchers` to `JsonMatchers`
**BREAKING CHANGE**

Rename module to `JsonMatchers`. This resolves clashing with gems like
`oj` / `oj_mimic_json` that take control of the standard library's
`json` module. As a result, the file to require is now `json_matchers`,
instead of `json/matchers`.

This should leave clients largely unaffected, since their usage of the
gem is based around `require "json/matchers"` and using a globally
accessible `match_response_schema`.

The only real difference is that the gem name (`json-matchers`) is no
longer aligned with Ruby conventions around module nesting. As a
consumer, I would expect the gem name to be `json_matchers`.

For backwards compatibility, this PR introduces the `json/matchers` file
as an alias for `json_matchers`, so that consumers can still call
`require "json/matchers"`.

For compatibility with the gem name, consumers can also call `require
"json-matchers"`, which is aliased to `require "json_matchers"`.

No longer condone auto-loading RSpec. Add documentation around adding
`require "json_matchers/rspec"` to consumers' `spec/spec_helper.rb`
