+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 master
 ------
 
+* BREAKING CHANGE: Rename module to `JsonMatchers`. This resolves clashing with
+  gems like `oj` / `oj_mimic_json` that take control of the standard library's
+  `json` module. As a result, the file to require is now `json_matchers`,
+  instead of `json/matchers`.
+
+* No longer condone auto-loading RSpec. Add documentation around adding `require
+  "json_matchers/rspec"` to consumers' `spec/spec_helper.rb`
+
 0.3.0
 -----
