--- lib/json/matchers/errors.rb
@@ -1,6 +1,0 @@
-module JSON
-  module Matchers
-    InvalidSchemaError = Class.new(StandardError)
-    MissingSchema = Class.new(Errno::ENOENT)
-  end
-end
