--- lib/json/matchers/version.rb
@@ -1,5 +1,0 @@
-module JSON
-  module Matchers
-    VERSION = "0.3.0"
-  end
-end
