+++ lib/json_matchers/errors.rb
@@ -1,0 +1,4 @@
+module JsonMatchers
+  InvalidSchemaError = Class.new(StandardError)
+  MissingSchema = Class.new(Errno::ENOENT)
+end
