+++ lib/json_matchers/version.rb
@@ -1,0 +1,3 @@
+module JsonMatchers
+  VERSION = "0.4.0"
+end
